requirements.txt
fastapi
sqlalchemy
dotvenv
passlib
python-jose
uvicorn